put step: add input dir validation
PCI-2587
