+++ cogito/testdata/a-repo/dot.git/HEAD.template
