+++ cogito/testdata/a-repo/dot.git/config.template
