+++ cogito/testdata/a-repo/dot.git/refs/heads/{{.branch_name}}.template
